refactor: organizing files in folders
